refactor(core): organize ai model root modules
