+++ packages/core/src/ai-model/workflows/planning/index.ts
@@ -1,2 +1,6 @@
+export {
+  ConversationHistory,
+  type ConversationHistoryOptions,
+} from './conversation-history';
 export { standardPlan } from './standard-planning';
 export { parseXMLPlanningResponse } from './standard-planning-parser';
