+++ packages/core/src/ai-model/workflows/planning/standard-planning.ts
@@ -8,7 +8,6 @@ import { getDebug } from '@midscene/shared/logger';
 import { assert } from '@midscene/shared/utils';
 import type { ChatCompletionMessageParam } from 'openai/resources/index';
 import { buildYamlFlowFromPlans } from '../../../common';
-import { planningModelFamilyRequiredForLocateMessage } from '../../errors';
 import { prepareModelImage } from '../../model-adapter/image-preprocess';
 import { systemPromptToTaskPlanning } from '../../prompt/llm-planning';
 import { AIResponseParseError, callAI } from '../../service-caller/index';
@@ -20,6 +19,7 @@ import type {
   LocateResultAdapter,
   LocateResultContext,
 } from '../../shared/model-locate-result';
+import { planningModelFamilyRequiredForLocateMessage } from '../../shared/model-locate-result/errors';
 import { normalizePlanningActionLocateFields } from './locate-normalization';
 import { parseXMLPlanningResponse } from './standard-planning-parser';
 import type { PlanOptions } from './types';
