+++ packages/core/tests/unit-test/custom-planning.test.ts
@@ -1,4 +1,4 @@
-import { ConversationHistory } from '@/ai-model/conversation-history';
+import { ConversationHistory } from '@/ai-model/workflows/planning/conversation-history';
 import { buildCustomPlanningMessages } from '@/ai-model/workflows/planning/custom-planning';
 import type { PlanOptions } from '@/ai-model/workflows/planning/types';
 import { describe, expect, it } from 'vitest';
